pyproject.toml
[project]
name = "study-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.7",
    "google-genai>=1.53.0",
    "langchain-text-splitters>=1.0.0",
    "pinecone>=8.0.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.6",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.21",
    "uvicorn>=0.38.0",
]
